INTERNAL: Remove duplicated logic in concrete classes that extends BTreeGetBulkImpl.
