+++ src/main/java/net/spy/memcached/collection/BTreeGetBulkWithLongTypeBkey.java
@@ -19,7 +19,6 @@ package net.spy.memcached.collection;
 import java.util.List;
 
 import net.spy.memcached.MemcachedNode;
-import net.spy.memcached.util.BTreeUtil;
 
 public class BTreeGetBulkWithLongTypeBkey<T> extends BTreeGetBulkImpl<T> {
 
@@ -34,29 +33,7 @@ public class BTreeGetBulkWithLongTypeBkey<T> extends BTreeGetBulkImpl<T> {
     return (Long) subkey;
   }
 
-  public void decodeItemHeader(String itemHeader) {
-    String[] splited = itemHeader.split(" ");
-
-    if (splited.length == 3) {
-      // ELEMENT <bkey> <bytes>
-      this.subkey = Long.parseLong(splited[1]);
-      this.dataLength = Integer.parseInt(splited[2]);
-      this.eflag = null;
-    } else if (splited.length == 4) {
-      // ELEMENT <bkey> <eflag> <bytes>
-      this.subkey = Long.parseLong(splited[1]);
-      this.eflag = BTreeUtil.hexStringToByteArrays(splited[2].substring(2));
-      this.dataLength = Integer.parseInt(splited[3]);
-    }
+  protected Object decodeSubkey(String subkey) {
+    return Long.parseLong(subkey);
   }
-
-  @Override
-  public void decodeKeyHeader(String keyHeader) {
-    String[] splited = keyHeader.split(" ");
-    this.key = splited[1];
-    if (splited.length == 5) {
-      this.flag = Integer.valueOf(splited[3]);
-    }
-  }
-
 }
